AWS/requirements.txt
paramiko==3.4.0
scp==0.14.5
mysql-connector-python==8.2.0
